Fix sim thread util regression (#488)
## 📝 Summary

This fixes a perf regression that we added when we started to check
provider factory health on each check. The core part of this PR is
https://github.com/flashbots/rbuilder/commit/ae8177ee1d63db5ca6350ab92748c0c481c9445c
since its directly fixed problem at hand: simulation thread utilization
is up from 2% to 20%.

This affects simulation the most since we would open a db transaction
for each order that we simulate.
It does not affect block building that much since we open a db
transaction per block building attempt instead of per block.

I've added
[this](https://github.com/flashbots/rbuilder/commit/54827e74e2a3610dc29377c109143815c74dfebf)
commit with more changes to make builders more efficient at opening db
transactions. Now we do it once per block instead of once per block
building attempt.

There are no other places in the code that do this reopening check in a
hot loop. The only thing that uses it like wallet balance watcher are
doing this per block. The exception is the code that checks last block
in a loop to cancel slots but its not critical to performance.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)